safeplan(S10.9): `SVAAFrameInfo` is `Sync` — the identity pair becomes `usize`, the array pair becomes slices
339 -> 339; raw_ptr 627 -> 619. `SVAAFrameInfo` 520 -> **504**,
`SVAAFrameInfoExt` 1536 -> **1520**, `SComplexityAnalysisParam` 40 -> **24**.

Four raw fields, two different reasons, two different fixes.

**`SVAACalcResult::pCurY` / `pRefY`: `*mut u8` -> `usize`.** The whole-tree read
is *one* comparison, in `adaptive_quantization.rs`: "reuse the VAA statistics when
they were computed over exactly this pair of pictures". **Nothing dereferences
them** — the walk that reads pixels takes its own cursors. They are identity
tokens, and storing an identity as a pointer bought nothing while costing
`SVAAFrameInfo` a `!Sync` reason. An address is plain `Copy` data, which is the
tree's own phrase from the two-thread probes that carry layer addresses across a
spawn the same way. Zero bytes moved.

**`SComplexityAnalysisParam::pBackgroundMbFlag` / `uiRefMbType`: slices at the
`Process` call.** This is T9.X's move for the two GOM arrays and Phase 6 session
B's for `pCalcResult`, applied to the last pair — the precedent was written down
in the struct's own doc.

**And that doc's reason for keeping them had expired.** It said the first is "a
self-pointer into `SVAAFrameInfo::pVaaBackgroundMbFlag`" and the second "has **no
writer in this port at all** — upstream fills it through `SetRefMbType`, which was
never ported". `SetRefMbType` *is* ported now, and does aim the pointer at a
reference picture's array. A self-pointer into a field of the enclosing struct is
a reason to *remove* a field, not to keep it.

`SetRefMbType` returns `Option<RecPicId>` rather than writing through a
`*mut *mut u32` out-parameter: an identity carries the same decision and lets the
caller take the slice under a borrow the compiler can see. The caller needs the
VAA block, the rate controller and the reference list at once, so
`vaa_rc_and_ref_list_mut` joins the §4.6 family — three owners, one call.

The three size pins move with the removals, as their own note prescribes: they are
**drift trackers, not ABI contracts** ("this struct has been off the C++'s 264
since Phase 6 session B and `repr(C)` came off with T6.F3"). All three
re-measured.

**One root left before `sWelsEncCtx: Sync`:**

    SLogContext  *mut c_void  (TraceUserCtx)

— an opaque C-callback user pointer with nothing to convert it *to*. That one is a
ruling (an audited `Sync` wrapper) rather than a conversion, and it is the last
thing between the tree and `SliceJobHandle` carrying a reference.

Referee: `cargo test --lib` 405/0; differential sweep **583/583** across all eight
presets including `bg`, which is this path's byte referee, byte-identical; ratchet
24 raw pointers down, no increases; span scan 0.

+++ rust/crates/openh264-rs/src/processing/vaacalc.rs
@@ -466,8 +466,8 @@ impl CVAACalculation {
             return RET_INVALIDPARAM;
         }
 
-        result.pCurY = pCurData;
-        result.pRefY = pRefData;
+        result.pCurY = pCurData as usize;
+        result.pRefY = pRefData as usize;
 
         // **T6.F3**: the five `VAACalc*_c` shims stood here, each handed the six out
         // arrays as bare pointers and each rebuilding a slice over them with
